requirements.txt
transformers
opencv-python
Pillow
numpy
pandas
onnxruntime-gpu
huggingface_hub